Backend/requirements.txt
appdirs==1.4.3
asn1crypto==0.22.0
bcrypt==3.1.3
beautifulsoup4==4.5.3
bleach==2.0.0
cffi==1.10.0
click==6.7
codecov==2.0.9
colorama==0.3.7
cookies==2.2.1
coreapi==2.3.0
coreschema==0.0.4
coverage==4.4.1
cryptography==1.9
decorator==4.0.11
defusedxml==0.5.0
detox==0.11
dj-database-url==0.4.2
Django==1.9.3
django-allauth==0.31.0
django-cors-headers==2.0.2
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-filter==1.0.2
django-hosts==2.0
django-jsonfield==1.0.1
django-mailgun==0.9.1
django-mssql==1.8
django-notify-x==0.1.7
django-webpack-loader==0.2.4
djangorestframework==3.3.2
entrypoints==0.2.2
enum-compat==0.0.2
eventlet==0.21.0
Fabric==1.10.2
Flask==0.12.2
greenlet==0.4.12
gunicorn==19.7.1
html5lib==0.999999999
idna==2.5
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Markdown==2.6.8
MarkupSafe==1.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
notebook==5.0.0
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandocfilters==1.4.1
paramiko==2.2.1
pickleshare==0.7.4
Pillow==4.1.1
pluggy==0.4.0
prompt-toolkit==1.0.14
py==1.4.33
pyasn1==0.2.3
pycparser==2.18
pygments==2.2.0
PyJWT==1.4.2
PyMySQL==0.7.10
PyNaCl==1.1.2
pyparsing==2.2.0
PySocks==1.6.7
pytest==3.1.2
pytest-cov==2.5.1
python-dateutil==2.6.0
python-decouple==3.0
python3-openid==3.1.0
pytz==2017.2
pyzmq==16.0.2
qtconsole==4.3.0
recastai==3.0.0
requests==2.13.0
requests-oauthlib==0.8.0
responses==0.5.1
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.2.0
sqlparse==0.2.3
testpath==0.3
tornado==4.5.1
tox==2.7.0
traitlets==4.3.2
twilio==6.3.0
Unipath==1.1
uritemplate==3.0.0
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.3.0
widgetsnbextension==2.0.0
win-unicode-console==0.5
wit==4.2.0
